requirements.txt
docutils==0.12
docxtpl==0.2.2
Jinja2==2.8
lxml==3.6.1
MarkupSafe==0.23
Pillow==3.2.0
pockets==0.3
Pygments==2.1.3
python-docx==0.8.6
reportlab==3.3.0
six==1.10.0
Sphinx==1.2.3
sphinxcontrib-napoleon==0.5.2
